moving software around
