+++ code/run_production_server.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
 
 gunicorn -w 4 -b 0.0.0.0:5001 --timeout 3600 main:app
-sh ./continuous_metadata_update.sh >>update_metadata.log 2>&1
