+++ code/utilities/run_production_populate.sh
@@ -1,6 +1,3 @@
 #!/bin/bash
 
-
-cd utilities
 python3 ./search_dataset_metadata.py all
-cd ..
